requirements.txt
astroid==2.9.3
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12
deepl-translate==1.2.0
dnspython==2.2.0
eventlet==0.33.0
flake8==4.0.1
greenlet==1.1.2
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
six==1.16.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
websocket-client==1.3.1
wrapt==1.13.3
